break down main server, worker pool and socket ocnfig
